Add log priority 'none' to disable service
This renames the configuration `JOURNAL_LOGLEVEL` to `JOURNAL_PRIORITY`
(`JOURNAL_LOGLEVEL` is still available but deprecated). It also adds a
new option for priority: `none`, indicating the service to ignore any
log events of any type. If this option is given, the service will shut
down to save resources.

Change-type: minor